pyproject.toml
[project]
name = "caption-ai"
version = "0.1.0"
description = "Meeting caption listener + summarizer with pluggable LLM backends"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Gabriel Parra", email = "gabeparra@example.com"}
]
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
stt = [
    "faster-whisper>=0.10.0",
    "sounddevice>=0.4.6",
]
browser = [
    "playwright>=1.40.0",
]
local = []

[project.scripts]
caption-ai = "caption_ai.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

